pyproject.toml
[project]
name = "vit-explainability-approximated-lrp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "black>=25.12.0",
    "einops>=0.8.1",
    "matplotlib>=3.10.8",
    "numpy>=2.2.6",
    "opencv-python>=4.11.0.86",
    "pillow>=12.1.0",
    "ruff>=0.14.11",
    "timm>=1.0.24",
    "torch>=2.9.1",
    "torchvision>=0.24.1",
    "transformers>=4.57.5",
]
